articles app 삭제, postings Comment 생성
